% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recmap.R
\name{as.recmap.SpatialPolygonsDataFrame}
\alias{as.recmap.SpatialPolygonsDataFrame}
\alias{as.recmap}
\title{Convert a SpatialPolygonsDataFrame Object to recmap Object}
\usage{
\method{as.recmap}{SpatialPolygonsDataFrame}(X)
}
\arguments{
\item{X}{\code{\link[sp]{SpatialPolygonsDataFrame}} object.}
}
\value{
returns a \code{\link[recmap]{recmap}} object.
}
\description{
The method generates a recmap class out of a \code{\link[sp]{SpatialPolygonsDataFrame}} object.
}
\examples{
checkerboard(8) |> recmap() |> as.SpatialPolygonsDataFrame() -> SpDf 
summary(SpDf)
spplot(SpDf)
summary(as.recmap(SpDf))

}
\references{
Roger S. Bivand, Edzer Pebesma, Virgilio Gomez-Rubio, 2013.
Applied spatial data analysis with R, Second edition. Springer, NY.
}
