% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recmap.R
\name{.draw_recmap_us_state_ev}
\alias{.draw_recmap_us_state_ev}
\title{this function reproduces the original election cartogram from 2004 using 
the cartogram output from the 2003 implementation.}
\usage{
.draw_recmap_us_state_ev(plot = TRUE)
}
\arguments{
\item{plot}{default is TRUE}
}
\value{
the plot
}
\description{
this function reproduces the original election cartogram from 2004 using 
the cartogram output from the 2003 implementation.
}
