% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recmap.R
\name{.getUS_map}
\alias{.getUS_map}
\title{return a overlapping minimal bounding boxes (MBB) of the U.S.
using \link[datasets]{datasets}.}
\usage{
.getUS_map()
}
\description{
return a overlapping minimal bounding boxes (MBB) of the U.S.
using \link[datasets]{datasets}.
}
\references{
\itemize{
\item \doi{10.18637/jss.v086.c01}
\item \doi{10.1109/INFVIS.2004.57}
}
}
\author{
Christian Panse, 2016
}
