% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recmap.R
\name{as.SpatialPolygonsDataFrame.recmap}
\alias{as.SpatialPolygonsDataFrame.recmap}
\alias{as.SpatialPolygonsDataFrame}
\title{Convert a recmap Object to SpatialPolygonsDataFrame Object.}
\usage{
\method{as.SpatialPolygonsDataFrame}{recmap}(x, df = NULL, ...)
}
\arguments{
\item{x}{a \code{\link{recmap}} object.}

\item{df}{a \code{data.frame} object. default is NULL.}

\item{\dots}{\dots}
}
\description{
The method generates a SpatialPolygons object of a as input given
\code{\link{recmap}} object. Both \code{data.frame}s are merged by the index order.
}
\examples{
SpDf <- as.SpatialPolygonsDataFrame(recmap(checkerboard(8)))
summary(SpDf)
spplot(SpDf)

}
