% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recmap.R
\name{plot.recmap}
\alias{plot.recmap}
\alias{plot.recmapGA}
\alias{plot.recmapGRASP}
\title{Plot a recmap object.}
\usage{
\method{plot}{recmap}(x, col = "#00000011", col.text = "grey", border = "darkgreen", ...)
}
\arguments{
\item{x}{\code{recmap} object - can be input or output of \code{recmap}.}

\item{col}{a vector of colors.}

\item{col.text}{a vector of colors.}

\item{border}{This parameter is passed to the \code{\link{rect}} function.
color for rectangle border(s). The default means par("fg"). 
Use border = NA to omit borders. If there are shading lines, border = TRUE 
means use the same colour for the border as for the shading lines.
The default value is set to \code{'darkgreen'}.}

\item{\ldots}{whatsoever}
}
\value{
graphical output
}
\description{
plots input and output of the \code{\link{recmap}} function.
The function requires column names (x, y, dx, dy).
}
\examples{
checkerboard(2) |> recmap() |> plot()
}
