% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recmap.R
\name{summary.recmap}
\alias{summary.recmap}
\alias{summary.recmapGA}
\title{Summary for recmap object}
\usage{
\method{summary}{recmap}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
returns a \code{data.frame} containing summary information, e.g.,
objective functions or number of map regions.
}
\description{
Summary method for S3 class \code{\link{recmap}}.
The area error is computed as described in the CartoDraw paper.
}
\examples{
summary(checkerboard(4));
summary(recmap(checkerboard(4)))
}
\references{
\doi{10.1109/TVCG.2004.1260761}
}
